reverting the commit 'extract internal/diskencryption package' so common solution would be used in the whole project.
a separate follow-up task will be added
