+++ internal/cluster/validations/validations.go
@@ -17,7 +17,6 @@ import (
 	"github.com/hashicorp/go-multierror"
 	"github.com/openshift/assisted-image-service/pkg/isoeditor"
 	"github.com/openshift/assisted-service/internal/common"
-	"github.com/openshift/assisted-service/internal/diskencryption"
 	"github.com/openshift/assisted-service/internal/featuresupport"
 	"github.com/openshift/assisted-service/internal/network"
 	"github.com/openshift/assisted-service/models"
@@ -777,7 +776,7 @@ func ValidateDiskEncryptionParams(diskEncryptionParams *models.DiskEncryption, D
 	if diskEncryptionParams == nil {
 		return nil
 	}
-	if !DiskEncryptionSupport && diskencryption.RequestsConfiguration(diskEncryptionParams) {
+	if !DiskEncryptionSupport && common.RequestsConfiguration(diskEncryptionParams) {
 		return errors.New("Disk encryption support is not enabled. Cannot apply configurations to the cluster")
 	}
 	if diskEncryptionParams.Mode != nil && swag.StringValue(diskEncryptionParams.Mode) == models.DiskEncryptionModeTang {
