+++ internal/common/disk_encryption.go
@@ -1,4 +1,4 @@
-package diskencryption
+package common
 
 import (
 	"strings"
