+++ internal/host/hostcommands/tang_connectivity_check_cmd.go
@@ -7,7 +7,6 @@ import (
 	ignition_types "github.com/coreos/ignition/v2/config/v3_2/types"
 	"github.com/go-openapi/swag"
 	"github.com/openshift/assisted-service/internal/common"
-	"github.com/openshift/assisted-service/internal/diskencryption"
 	"github.com/openshift/assisted-service/internal/host/hostutil"
 	"github.com/openshift/assisted-service/models"
 	"github.com/sirupsen/logrus"
@@ -71,9 +70,9 @@ func (c *tangConnectivityCheckCmd) shouldRunTangConnectivityCheck(cluster common
 	// 1. DiskEncryption not set or not enabled.
 	// 2. DiskEncryption mode is not tang based.
 	// 3. DiskEncryption is not enabled, for the host role.
-	if !diskencryption.IsConfigured(cluster.DiskEncryption) ||
+	if !common.IsConfigured(cluster.DiskEncryption) ||
 		swag.StringValue(cluster.DiskEncryption.Mode) == models.DiskEncryptionModeTpmv2 ||
-		!diskencryption.EnabledForRole(*cluster.DiskEncryption, common.GetEffectiveRole(host)) {
+		!common.EnabledForRole(*cluster.DiskEncryption, common.GetEffectiveRole(host)) {
 		c.log.Debugf("skipping tangConnectivityCheck for host %s, cluster DiskEncryption config does not require validation here",
 			host.ID.String())
 		return false
