+++ internal/host/validator.go
@@ -17,7 +17,6 @@ import (
 	"github.com/go-openapi/swag"
 	"github.com/openshift/assisted-service/internal/common"
 	"github.com/openshift/assisted-service/internal/constants"
-	"github.com/openshift/assisted-service/internal/diskencryption"
 	"github.com/openshift/assisted-service/internal/hardware"
 	"github.com/openshift/assisted-service/internal/host/hostutil"
 	"github.com/openshift/assisted-service/internal/network"
@@ -492,7 +491,7 @@ func (v *validator) diskEncryptionRequirementsSatisfied(c *validationContext) (V
 	if c.infraEnv != nil {
 		return ValidationSuccessSuppressOutput, ""
 	}
-	if !hostutil.IsDay2Host(c.host) && !diskencryption.IsConfigured(c.cluster.DiskEncryption) {
+	if !hostutil.IsDay2Host(c.host) && !common.IsConfigured(c.cluster.DiskEncryption) {
 		return ValidationSuccessSuppressOutput, ""
 	}
 	if c.inventory == nil {
@@ -536,7 +535,7 @@ func (v *validator) diskEncryptionRequirementsSatisfied(c *validationContext) (V
 		if role == models.HostRoleAutoAssign {
 			return ValidationPending, "Missing role assignment"
 		}
-		if !diskencryption.EnabledForRole(*c.cluster.DiskEncryption, role) {
+		if !common.EnabledForRole(*c.cluster.DiskEncryption, role) {
 			return ValidationSuccessSuppressOutput, ""
 		}
 		if swag.StringValue(c.cluster.DiskEncryption.Mode) == models.DiskEncryptionModeTang {
