+++ internal/network/manifests_generator.go
@@ -12,7 +12,6 @@ import (
 	"github.com/go-openapi/swag"
 	"github.com/kelseyhightower/envconfig"
 	"github.com/openshift/assisted-service/internal/common"
-	"github.com/openshift/assisted-service/internal/diskencryption"
 	manifestsapi "github.com/openshift/assisted-service/internal/manifests/api"
 	operatorcommon "github.com/openshift/assisted-service/internal/operators/common"
 	"github.com/openshift/assisted-service/internal/operators/openshiftai"
@@ -350,7 +349,7 @@ func (m *ManifestsGenerator) createDiskEncryptionManifest(ctx context.Context, l
 
 func (m *ManifestsGenerator) AddDiskEncryptionManifest(ctx context.Context, log logrus.FieldLogger, c *common.Cluster) error {
 
-	if !diskencryption.IsConfigured(c.DiskEncryption) {
+	if !common.IsConfigured(c.DiskEncryption) {
 		return nil
 	}
